+++ server/app.js
@@ -2,7 +2,7 @@ const express = require('express')
 const app = express()
 const cors = require('cors')
 const port = process.env.PORT || 3000
-const { getRecipes, getUserFavouritedRecipes, getLoggedInUserRecipes } = require('./controller/recipeController')
+const { getRecipes, getUserFavouritedRecipes, getLoggedInUserRecipes, getRecipeDetail } = require('./controller/recipeController')
 
 app.use(express.urlencoded({extended: true}))
 app.use(express.json())
@@ -26,10 +26,12 @@ app.post('/users', instanceMulter.single('imageFile'), imageKitUpload,  (req, re
 
 // fetch all recipes 
 app.get('/recipes', getRecipes)
-// fetch all user favourited recipes
+// fetch all user favourited recipes (untested)
 app.get('/recipes/favourite', getUserFavouritedRecipes)
-// fetch all logged in user recipes
+// fetch all logged in user recipes (untested)
 app.get('/recipes/own', getLoggedInUserRecipes)
+// fetch recipe detail
+app.get('/recipes/:id', getRecipeDetail)
 
 app.listen(port, () => {
     console.log(`app listening to http://localhost:${port}`)
